fix: CsvReader[(A, B)] implicit can not be found
If CsvReader is declared to be covariant test does not compile because implicit resolution can not work with implicit production rule and an invariant CsvColumnReader.
This can be fixed by declaring CsvColumnReader as covariant, but it happens covariance has complex intricacies. One could declare use a reader for a part of an inheritance tree and claiming he has a reader for the whole tree. By making reader invariant the compiler will let the user know he can claim such fact so the decision of how to handle this issue is delegated to the user.

Rollback to an invariant CsvReader[A]